add "open with" dialogue box in ipad.
This contains the fix for commit - https://github.com/kristikristo/react-native-file-opener/commit/89ba83ed63776d51030d8c6d423fa48b22cd5aed

CGRectMake in line 14 is not a proper return type. It is a function. hence it is removed and the menu containing the list of applications to open the file with is created manually in line 37.

fixes huangzuizui/react-native-file-opener#15